refactor: remove mobile directory
